+++ examples/repl/google_colab_sandbox_example.py
@@ -5,9 +5,7 @@
 
 import os
 
-from repl_common import run_repl
-
-from code_sandboxes import Sandbox
+from code_sandboxes import Sandbox, run_repl
 
 
 def _require(name: str) -> str: